Change the leader and cluster discovery
Reflect changes in etcd: the format of the cluster introspection features now includes "http://" and the leader and machine URIs have become versioned.

This doesn't work with the current version of etcd (v0.1.1) since it returns the wrong port for the leader URI (it returns the Raft port, not the Etcd port).